requirements.txt
python-dotenv==1.0.0
requests==2.31.0
paramiko==3.4.0
